+++ src/config.php
@@ -27,6 +27,7 @@ return array(
     'cacheBar' => true,
     'displayGreeting' => true,
     'displayDashboardLink' => true,
+    'displayDefaultEditSection' => true,
     'displaySettingsLink' => true,
     'displayLogout' => true,
     'enableMobileMenu' => true,